+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Unreleased
 
+- Added `_v` versions of `call` methods to make it easier to pass commands as arrays without splating.
 - Fix calling `blocking_call` with a block in a pipeline.
 - `blocking_call` now raise `ReadTimeoutError` if the command didn't complete in time.
 - Fix `blocking_call` to not respect `retry_attempts` on timeout.